requirements.txt
pywebview; sys_platform == 'win32'
pywebview[qt]; sys_platform == 'darwin'
pywebview[qt]; sys_platform == 'linux'
pywebview[qt]; sys_platform == 'linux2'
requests
Flask
pydantic